+++ pybamm/expression_tree/operations/evaluate.py
@@ -175,6 +175,9 @@ def find_symbols(symbol, constant_symbols, variable_symbols):
     elif isinstance(symbol, pybamm.Time):
         symbol_str = "t"
 
+    elif isinstance(symbol, pybamm.InputParameter):
+        symbol_str = "u['{}']".format(symbol.name)
+
     else:
         raise NotImplementedError(
             "Not implemented for a symbol of type '{}'".format(type(symbol))
@@ -270,7 +273,7 @@ class EvaluatorPython:
             self._result_var, "return" + self._result_var, "eval"
         )
 
-    def evaluate(self, t=None, y=None, known_evals=None):
+    def evaluate(self, t=None, y=None, u=None, known_evals=None):
         """
         Acts as a drop-in replacement for :func:`pybamm.Symbol.evaluate`
         """