+++ pybamm/solvers/base_solver.py
@@ -49,7 +49,7 @@ class BaseSolver(object):
     def atol(self, value):
         self._atol = value
 
-    def solve(self, model, t_eval):
+    def solve(self, model, t_eval, inputs=None):
         """
         Execute the solver setup and calculate the solution of the model at
         specified times.
@@ -61,6 +61,8 @@ class BaseSolver(object):
             initial_conditions
         t_eval : numeric type
             The times at which to compute the solution
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
 
         Raises
         ------
@@ -77,14 +79,17 @@ class BaseSolver(object):
         # Set up
         timer = pybamm.Timer()
         start_time = timer.time()
+        inputs = inputs or {}
         if model.convert_to_format == "casadi" or isinstance(self, pybamm.CasadiSolver):
-            self.set_up_casadi(model)
+            self.set_up_casadi(model, inputs)
         else:
-            self.set_up(model)
+            self.set_up(model, inputs)
         set_up_time = timer.time() - start_time
 
         # Solve
-        solution, solve_time, termination = self.compute_solution(model, t_eval)
+        solution, solve_time, termination = self.compute_solution(
+            model, t_eval, inputs=inputs
+        )
 
         # Assign times
         solution.solve_time = solve_time
@@ -100,7 +105,7 @@ class BaseSolver(object):
         )
         return solution
 
-    def step(self, model, dt, npts=2, log=True, external_variables=None):
+    def step(self, model, dt, npts=2, log=True, external_variables=None, inputs=None):
         """
         Step the solution of the model forward by a given time increment. The
         first time this method is called it executes the necessary setup by
@@ -119,6 +124,9 @@ class BaseSolver(object):
         external_variables : dict
             A dictionary of external variables and their corresponding
             values at the current time
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
+
 
         Raises
         ------
@@ -132,6 +140,7 @@ class BaseSolver(object):
 
         # Set timer
         timer = pybamm.Timer()
+        inputs = inputs or {}
 
         if not hasattr(self, "y0"):
             # create a y_pad vector of the correct size:
@@ -148,12 +157,12 @@ class BaseSolver(object):
             if model.convert_to_format == "casadi" or isinstance(
                 self, pybamm.CasadiSolver
             ):
-                self.set_up_casadi(model)
+                self.set_up_casadi(model, inputs)
             else:
                 pybamm.logger.debug(
                     "Start stepping {} with {}".format(model.name, self.name)
                 )
-                self.set_up(model)
+                self.set_up(model, inputs)
             self.t = 0.0
             set_up_time = timer.time()
 
@@ -162,7 +171,7 @@ class BaseSolver(object):
 
         # Step
         t_eval = np.linspace(self.t, self.t + dt, npts)
-        solution, solve_time, termination = self.compute_solution(model, t_eval)
+        solution, solve_time, termination = self.compute_solution(model, t_eval, inputs)
 
         # Set self.t and self.y0 to their values at the final step
         self.t = solution.t[-1]
@@ -221,7 +230,7 @@ class BaseSolver(object):
                 )
             self.y_ext[y_slice] = var_vals
 
-    def compute_solution(self, model, t_eval):
+    def compute_solution(self, model, t_eval, inputs=None):
         """Calculate the solution of the model at specified times. Note: this
         does *not* execute the solver setup.
 
@@ -232,11 +241,13 @@ class BaseSolver(object):
             initial_conditions
         t_eval : numeric type
             The times at which to compute the solution
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
 
         """
         raise NotImplementedError
 
-    def set_up(self, model):
+    def set_up(self, model, inputs=None):
         """Unpack model, perform checks, simplify and calculate jacobian.
 
         Parameters
@@ -244,11 +255,13 @@ class BaseSolver(object):
         model : :class:`pybamm.BaseModel`
             The model whose solution to calculate. Must have attributes rhs and
             initial_conditions
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
 
         """
         raise NotImplementedError
 
-    def set_up_casadi(self, model):
+    def set_up_casadi(self, model, inputs=None):
         """Convert model to casadi format and use their inbuilt functionalities.
 
         Parameters
@@ -256,6 +269,8 @@ class BaseSolver(object):
         model : :class:`pybamm.BaseModel`
             The model whose solution to calculate. Must have attributes rhs and
             initial_conditions
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
 
         """
         raise NotImplementedError